Python_tracking_4_with_gps/requirements.txt

flask 
pyserial 
pandas  
openpyxl 
Flask-SocketIO
geopy
haversine      
python-dotenv
eventlet       